Create a new class for Popups
